Make uidvalidity field 64-bit
But not in the summary yet; I'll get to that shortly and we aren't even
storing it there anyway right now -- we're always writing zero.
(cherry picked from commit 8cd40b296bfe1bdb9a7b46b8585a27872990af55)

+++ camel/providers/imapx/camel-imapx-server.h
@@ -79,7 +79,7 @@ struct _CamelIMAPXServer {
 	struct _CamelFolderChangeInfo *changes;
 	struct _CamelFolder *select_pending;
 	guint32 permanentflags;
-	guint32 uidvalidity;
+	guint64 uidvalidity;
 	guint32 unseen;
 	guint32 exists;
 	guint32 recent;
